@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=3.10.0"
 
 dependencies:
   flutter:
@@ -27,32 +27,33 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
+  cupertino_icons: ^1.0.8
 
   # sqlite package for managing the database
-  sqflite: ^1.3.2+2
+  sqflite: ^2.4.2
 
   # tools for path manipulation of filename strings
-  path: ^1.7.0
+  path: ^1.9.1
 
   # tool for getting commonly used filesystem locations
   path_provider: ^2.0.1
 
   # tool for traversing the folder structure
-  flutter_file_manager: ^0.2.0
+  #flutter_file_manager: ^0.2.0
+  file_picker: ^10.3.6
 
   # for managing permissions
-  permission_handler: ^6.1.1
+  permission_handler: ^12.0.1
 
   # for choosing colors
-  flex_color_picker: ^2.0.1
+  flex_color_picker: ^3.7.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
   # Change the launcher icon
-  flutter_launcher_icons: ^0.9.0
+  flutter_launcher_icons: ^0.14.4
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
